Refactored AbilityHandler
There is now a parent AbilityHandler class that just deals with activating abilities.
Player now uses a PlayerAbilityHandler : AbilityHandler
Turret now uses TurretHandler : AbilityHandler

Since the player didnt rotate anymore I added a "AbilityRotation" GameObject to the PlayerPrefab. Its rotation gets updated in PlayerMovement script in funktion LookAt.
This AbilityRotation GameObject is used to make AbilityHandler not rely on where the mouse is, that will be important when other objects want to cast a ability that dont have a mouse.